Fixing how form control is passed between parent form group and nested form groups to solve validation issues.
I was passing the nested FormGroup to each component that belonged to that form group as an @Input, then using this to
set the [formGroup] label on the 'parent' html containers for each individual field. Within those containers I was then
setting the formControlName for the individual fields. This was causing error states to propagate between sibling
field containers, but not between the fields themselves. So it was working for form submission when all of the fields were
valid, but when I started applying validation styles to individual fields the containers for each child of a nested
group were inheriting the error state from other fields in that group and causing the styles being displayed to
correspond to the group as a whole rather than the individual fields.

Fixed this by passing FormControl as @Input to the child components instead of FormGroup, and then removing the
FormGroup label from the field's 'parent' containter. Then I changed the formControlName label on the field
itself to a [formControl] label and set this to the name of the control that was passed from the parent component.
Also had to change getters in parent component to get the individual FormControl instead of the FormGroup. These
changes have only been implemented on serverRegion, gameEnvironment, and gameExperience. Will be making the same
changes to the other fields next.