Finishing create listing form validation
Changing the rest of the form fields to pass FormControl as @Input instead of FormGroup.
Adding styling for 'invalid' fields.
Adding in-line error messages for invalid fields.
Creating custom validator for group status and its corresponding event schedule fields
so that eventScheduleDate, time, and time zone become required if groupStatus is set to
'Future/Scheduled'. Adding method that checks for invalid 'requried' states for any of
those three fields and displays an error message for all three fields if any of them are
invalid.
Increasing font size for text input fields in create listing form.
Adding bindLabel/Value for time zone so that it displays and/or binds the correct value.