Moving group listing modal to its own component to separate from group-listings table
creating new component 'group-listing-modal'. Moving HTML and CSS for modal from group-listings component to new modal component.
Modifying instructions for row click to trigger onRowClick method in groupListing component which assigns tempListing to value:
selectedListing and passes this to the modal component as an @Input. OnRowClick also sets isModalVisible to true.
Adding closeModal method and closeOnBackdropClick method to modal ts to replicate previous close functionalities. OnModalClose
in parent component is triggered by event emitter in modal component when the modal is closed and sets visibility boolean to false
and selectedListing to null.