Writing unit tests for gameplay category and subcategory services. Fixing issues with convert to entity methods
I was trying to do the conversion from dto to entity for these fields incorrectly
by using a model mapper to map between the dto's and entities when I should have
just been querying the repositories for entities with matching id's and returning
those.

Modified the convertToEntity methods for category and subcategory services to
find the existing entity by id and verify that the dto/entity names match and
then return the existing entity. If either the id is not found or the name
does not match then an exception is thrown.

Wrote unit tests for happy cases of these conversions, id not found, or name
mismatch.