Changing all primitive types for entities, dtos, and corresponding arguments to wrapper classes
While writing unit tests for createGroupListing validators I discovered that the primitive
types default to undesirable values when a field is empty. This was causing issues in my
validation tests where empty values were defaulting to valid values because prmitives
like int defaulting to 0 allows them to pass 'not null' validation when they should not.
I decided that I needed to fix this because it could/would cause incorrect behavior for
empty fields during groupListing creation or other situations where it would cause invalid
entities to default to valid values.