Fixing error with nullable fields for listing creation preventing successful creation
After changing the data types from primitives to wrapper classes for entity id fields,
the 'not required' fields in the create lising form were throwing errors if left blank.
The explicit dto -> entity model mapper was causing hibernate to try to create instances
of the fields that had null id's which was throwing the error. I added conditional logic
to the model mapper for nullable fields to check whether they are null, if they are null
model mapper just returns null for that field rather than trying to map a null object.
If the field is not null then it maps accordingly.

I also added annotations to grouplisting entity and createlisting dto for @Nullable and
corrected some discrepencies between subcategory validation.

I changed the data type for timestamps to be of type Instant rather than localDateTime
to keep consistent with UTC in the database and time conversions on the front-end based
on user system default time zone.