Resolving grouplisting response dto issue with nullable fields
The data type change from primitives to wrapper classes also caused errors with
the response dtos on the front end with various operations being unable to be
performed on 'null' data.

I created another mapper config for the responseDto to explicitly map the entity
to dto. The mapper handles fields that are nullable with a conditional check for
null that changes the field to an empty string for the dto if the entity value
was null. I replaced the previous default model mapper in the group listing
service impl with the explicit mapper.

I changed the data type for timestamps in the response dto to align with the
global UTC time for data that will then be converted to user time on the front
end.

I set the global and local time to UTC in the database.

I also removed unnecessary referential actions ('on delete ...') from database
keys and adjusted NOT NULL constraints to match recent changes on entities/dtos.

Back to what I was working on... I corrected an assertion in the GroupListingService
test to match the validation change on 'subcategories'. Then I finished the test
for processing a successful group listing with all fields not null.