Moving getAll methods for environment and subcategory to separate service for caching
Spring cannot intercept queries for data that has been cached if the cacheable
method is in the same service as the query method. I moved the caching of the
getAll methods for these two services into a separate service so that the
cached data can be used for the getById methods to further reduce database
queries and minimize the amount of cached data.

To implement this I created a new service method for cacheAll in the caching
services and copied the existing getAll method body into the new service. Then
I just called the new cacheAll method for the getAll method.

For the findById I used a stream of the list generated by cacheAll to find and
return the dto with the queried id.

I also added the convertToDto methods to their respective interfaces so that they
can be accessed via interface reference in other classes. This was necessary for
caching the entities from the db query as dtos.

Finally, I refactored the unit tests for the subcategory and environment services
to account for the changes made to the getAll and findById methods. I also created
unit tests for the new caching services for these two classes that include
cacheAll_Found and cacheAll_NotFound.

I will be making these changes to the other lookup table classes as well to improve
caching.