Moving getAll for gameplayCategory service into separate service for caching Copied getAll method from GameplayCategoryServiceImpl into new caching service class to enable use of cached getAll() data in getById method.
Modified existing getAll and getById methods in GameplayCategoryServiceImpl to call the
CateogryCachingService.cacheAllExperiences().

Added convertToDto method to GameplayCategoryService interface and @override on the implementation of the method so that the convertToDto can be used in the separate caching service.

Wrote unit tests for CategoryCachingServiceImpl. Refactored tests in
GameplayCategoryServiceImplTest to reflect getAll and getById method changes.