Moving getAll for GroupStatusService into separate service for caching Copied getAll method from GroupStatusServiceImpl into new caching service class to enable use of cached getAll() data in getById method.
Modified existing getAll and getById methods in GroupStatusServiceImpl to call the
GroupStatusCachingService.cacheAllExperiences().

Added convertToDto method to GroupStatusService interface and @override on the implementation of the method so that the convertToDto can be used in the separate caching service.

Wrote unit tests for GroupStatusCachingServiceImpl. Refactored tests in
GroupStatusServiceImplTest to reflect getAll and getById method changes.