Moving getAll for LegalityService into separate service for caching Copied getAll method from LegalityServiceImpl into new caching service class to enable use of cached getAll() data in getById method.
Modified existing getAll and getById methods in LegalityServiceImpl to call the
LegalityCachingService.cacheAllExperiences().

Added convertToDto method to LegalityService interface and @override on the implementation of the method so that the convertToDto can be used in the separate caching service.

Wrote unit tests for LegalityCachingServiceImpl. Refactored tests in
LegalityServiceImplTest to reflect getAll and getById method changes.