FOUND ISSUE while working on moving getAll for PlanetMoonSystem into separate caching service
I am working on moving the getAll method from the planetMoonSystemSerice into a separate class
for a caching service so that the other getter methods can use the cached data. I was running
the app to test some things after writing unit tests and I encountered a circular dependncies
error. I think that this is because the caching services that I have been creating are injecting
the service classes so that the caching service can use the convertToDto methods, but the
service classes are also injecting the caching service to call the cacheAll methods. I probably
need to either move the convertToDto methods into the caching services, or possibly just create
separate services for the convertToDto methods, or a single covnertToDto service for all the
lookup tables.