moving conversion functions for GameEnivronment to separate service to resolve circular dependencies issue
Moved the convertToDto and convertToEntity methods from the GameEnvironmentService to new
GameEnvironmentConversionService. Moved the unit tests for these methods to a new test class
for GameEnvironmentConversionService. Refactored the caching service so that it calls
convertToDto from the new correct location.

The caching services for the lookup tables require the convertToDto method that was previously
in the GameEnvironmentServiceImpl. This was creating a circular dependency because the
GameEnvironmentServiceImpl required the cacheAll function from the caching service. I created
a separate conversion_services package for the conversion functions convertToDto and
convertToEntity in order to resolve this circular dependency issue.