moving conversion functions for GameplayCategory to new separate service to resolve circular dependencies issue
Moved the convertToDto and convertToEntity methods from the GameplayCategoryService to new
GameplayCategoryConversionService. Moved the unit tests for these methods to a new test class
for GameplayCategoryConversionService. Refactored the caching service so that it calls
convertToDto from the new correct location.

The caching services for the lookup tables require the convertToDto method that was previously
in the GameplayCategoryServiceImpl. This was creating a circular dependency because the
GameplayCategoryServiceImpl required the cacheAll function from the caching service. I created
a separate conversion_services package for the conversion functions convertToDto and
convertToEntity in order to resolve this circular dependency issue.