refactored PlayStyleService to separate caching methods, conversion methods, and crud methods into separate service classes Created a PlayStyleCachingService and implementation class to contain the getAll() logic for retrieving the playstyles from the repository and caching them with eh cache. Modified the existing PlayStyleService getAll() method to call the cacheAll() method, replacing the previous getAll() logic. Modified existing unit tests for PlayStyleService to reflect getAll/cacheAll changes. Moved the convertToDto and convertToEntity methods for PlayStyle to new separate PlayStyleConversionService and implementation to resolve circular dependencies issue. Wrote unit tests for these methods for success and failure paths. Refactored the caching service so that it calls convertToDto from the new correct location. Also added various logging statements for explicit validation failures defined in planetMoonSystem and Legality services.
Reasoning for Changes:
The caching for the lookup tables needed to be moved into its own service because Spring Caching creates a proxy around the bean that contains the caching, but a method called from within that same bean to the caching method bypasses the proxy.  Bypassing the proxy prevents the caching from actually taking place. So, the call to the caching method needs to take place from a separate bean, which is why I created a separate service class for the caching.
The caching services for the lookup tables require the convertToDto method that was previously in the PlayStyleServiceImpl. This was creating a circular dependency because the PlayStyleServiceImpl required the cacheAll function from the caching service. I created a separate conversion_services package for the conversion functions convertToDto and convertToEntity in order to resolve this circular dependency issue.