Moving GroupListing conversion services into separate service for consistency with other services
I also realized that I was misusing the .orElseGet function to throw exceptions and log errors.
I changed all incorrect usages of .orElseGet to .orElseThrow.

I was getting a warning when trying to use .orElseThrow with a log message and 'throw new'.  .orElseGet
was a solution that I think intellij just recommended, but I was not providing any backup/alternatives
with it. I realized while making some of the changes for groupListing conversion service that I
just needed to change 'throw new' to 'return new' in order to be able to use .orElseThrow that
produces an error log and the exception.