Configure backend to take jwt from request header for create listing and associate the listing creation with an existing user.
Remove @NotNull from user field in CreateGroupListingDto because the payload from the frontend will not have the userId in the dto yet.
In groupListingController, extract the keycloakId from the jwt in createGroupListing, find the associated user in the repository, and assign that userId to the listing DTO.
Rename UserResponseDto to 'PrivateUserResponseDto' because I need to create a public version with fewer details.
Update dev-realm to include display name for realm 'SC FleetFinder'.