Update userController logic to separate localUser check and create user into separate endpoints. Create separate public/private user response dtos and distinct conversion methods each.
At the time of authorization, frontend sends a 'get' request to /api/users/me to see if there is an existing user with the keycloakId from the jwt.
If a user exists, a PrivateUserResponseDto is returned for populating frontend's localUser$ for use in user profile/account info.
If a user does not exist, /me returns a 404 which causes the frontend to send a post request to /api/users/create-user to create
a new user in the app database using the claims from the jwt.
For separation of concerns and to ensure private data is only shared with the correct athorized user, PrivateUserResponseDtos are only accessible
via a keycloakId extracted from a jwt. PublicUserResponseDtos can be accessed with the app db userId but are still currently protected.

Wrote unit tests for convert to public/private dtos, and integration tests for UserController createUser, getUsers, getUserById.