Change EC2 deployment stage:
Pipeline is not pulling the most recent images from ECR and replacing the image: fields in the compose.yml with the previous strategy.
Trying a different approach with creating vars in the EC2 instance to grep the exact line for each image: for frontend and backend,
and then replace exactly that line with the new image tag from the previous stage.