Actually Fixed: change the image tag in compose.yml to new tag
Realized that one of these did actually successfully pull the image tag and replace that line in the docker compose,
but it failed for a different reason so i did not realize it at the time. Found the commit for that and copied the EOF syntax
into this. I think it failed because the closing 'EOF' was not in single quotes because it had an error that said something
along the lines of ''EOF' delineated by end of file'. I am fairly confident that this one is actually going to work, but
Keycloak is still not configured correctly for prod because it isnt sending JWTs for the /me queries.