Frontend changes to implement update listing value patching.
Move the patching for category, subcategory, system, and planet into separate block using setValue so that valueChanges subscriptions are triggered.
Change category and system dropdowns to use ID as bindValue so that id can be patched into dropdowns (and for consistency with other dropdowns).
Move the filter functionality for subcategories and planets based on parent selection into filter methods, and call filter in the fetch so that the
dropdowns properly update before patching.
Change create and update listing request objects/models to use the new bindValue from category and system dropdowns.
Add cancel buttons to the create listing and update listing forms (redirects to user-account).
Add object prints in user listing table on element select and on update listing clicks to print listing fields to console (in prod) for verification.