Fixes for update listing process:
Add control resets on subcategory and planets in the component subscriptions so the dropdowns reset on changes.
Fix comms option value in update-listing-request so that it passes the option not an object.
Change the data type in form service for comms option and comms service to string instead of <any>.