Profile styling and responsive design.
Changed isMobile breakpoint in user-acct-listings-table and mod-listings-table to 499 to catch mobile devices (was 399).
Added an options column to the user-acct table that is displayed when isMobile, replacing the select column with a menu.
Add conditional to only display user action buttons in table when display is not isMobile.
Add Delete confirmation check with mat-dialog pop-up.