Working filter: backend changes.
Add enum for comms option to match the id from the frontend to the enum options in the db (should not have done this to begin with.)
Add logic in the listingFIlter specification to add 'availableRoles' to the search input like criteria.
Remove the newly added label value from the frontend for populating filters bubbles from the filter string passed to the backend.
Needed this so that the filter bubbles on the front end could reference the filter element in the string for individual removal.
Added logic to map commsOption filters to db enums using the new commsOption enum and finding by id.