Frontend changes: add date range to filter options.
Add mat-date-range-input field to the searchbar component in the filters section as a template that appears when groupStatus and future/scheduled are selected.
Convert the date to a string YYYY-mm-dd on filter add (stored in the dto in this format).
Change the eventSchedule field to two separate fields in the interface and listingFilterRequest for start and end date.