Backend work for get/add/delete bookmarks.
fixed the bookmark repository so that it returns sets instead of single optionals.
Changed the user controller for delete bookmark to use a path variable with just the groupId, and then
get user from the jwt and add user and bookmarkid to the deletebookmark dto in the controller and service.