Frontend changes to add/delete/get bookmarks.
Created a user-bookmark service to move the http requests out of user-listing service and into their own service.
Added a behavior subject and observable of it to track changes to auth status and bookmarks for real-time responses to add/delete in ui.
Add solid bookmarked icon to group listing view table for when a listing is bookmarked by a user, and the outline wheen it is not.
Use isBookmarked method on row.groupId and async of the bookmarks observable from the bookmarks service to check for a bookmark.