Cleanup bookmark and report dtos and service methods to remove null user field from dto.
Removed unused user fields from add-bookmark-request and submit-listing-report on frontend.
Removed user fields from the corresponding backend dtos: AddBookmarkRequestDto and SubmitListingReportDto.
Changed the controller methods for both to just pass the user object as an argument.
Changed the service methods for add bookmark and submit listing to include Users as a paramater.
Removed the convertToEntity method from bookmarkconversionservice and just added a constructor to the class with parameters for listing and user.