Backend changes for moderation issue api endpoint and fixing the groupListingResponseDtoMapperConfig.
Create a moderation issue response dto and a model mapper config for converting the entity to the dto so it is not passing entire group objects and child object references.
Add a generic sortablePageRequestDto for page requests from sortable tables, but no other page criteria like filters.
Add a mod controller endpoint for get all mod issues with page and sort, and a corresponding service layer method for retrieving from repo.
Add a moderation issue conversion service for calling the convert from entity to dto from within the service on a page of issue entities.
FIX GROUP LISTING MAPPER CONFIG: add mappings for the id fields in groupListing because they were just being sent as null since I only mapped name/label values.
Replace the generic mapper in the groupListingConversion service convert to responseDto method with the adjusted modelmapper.
Add userId field to the groupListingResponseDto and mapping in the mapper.