Frontend and backend changes to fix time conversions.
Removed the frontend time conversions in the create and update listing requests.
Changed the frontend and backend dto fields for create and update to use three separate string fields for date, time, and tz.
Convert them on the backend in the model mapper to instants to store in the db (UTC).
Let the frontend do the automatic conversions for display, but no explicit conversions because that was not working the way I thought it was.